Autoauto Static Linking

+++ TeamCode/src/main/scripts/autoauto-compiler/compiler/transmutations/codebasewide-tasks/postprocess/zz-cleanup/index.js
@@ -1,14 +1,26 @@
+"use strict";
+
+const commandLineInterface = require("../../../../../../command-line-interface");
 const safeFsUtils = require("../../../../../../script-helpers/safe-fs-utils");
 const deleteOldGeneratedFiles = require("./delete-old-generated-files");
 
-module.exports = function (codebaseContext, contexts) {
-    var newFiles = Object.keys(codebaseContext.writtenFiles);
+/**
+ * @type {import("../../..").CodebaseTransmutateFunction}
+ */
+module.exports = function (context, contexts) {
     
-    for(const ctx of contexts) {
-        newFiles = newFiles.concat(Object.keys(ctx.writtenFiles));
-    }
+    if(commandLineInterface["run-cleanup"] === true) {
+        var newFiles = Object.keys(context.writtenFiles);
+        
+        for(const ctx of contexts) {
+            if(ctx.success === "SUCCESS") {
+                newFiles.push(...Object.keys(ctx.fileContext.writtenFiles));
+            }
+        }
+        
 
-    safeFsUtils.cleanDirectory(contexts[0].resultRoot, newFiles);
-    
-    deleteOldGeneratedFiles();
+        safeFsUtils.cleanDirectory(context.resultRoot, newFiles);
+        
+        deleteOldGeneratedFiles();
+    }
 }
